refactoring and new interface
