+++ app/src/main/java/com/frateinc/yopbooking2/ApiMethod/FindEventById.java
@@ -1,10 +1,10 @@
-package com.frateinc.yopbooking2;
+package com.frateinc.yopbooking2.ApiMethod;
 
 import android.os.AsyncTask;
 
-import com.frateinc.yopbooking2.models.Event;
+import com.frateinc.yopbooking2.Config.ConfigApi;
+import com.frateinc.yopbooking2.Models.Event;
 
-import org.json.JSONArray;
 import org.json.JSONObject;
 
 import java.io.BufferedInputStream;
@@ -16,12 +16,7 @@ import java.net.URL;
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.Iterator;
-import java.util.List;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 import javax.net.ssl.HttpsURLConnection;
 
@@ -29,20 +24,19 @@ import javax.net.ssl.HttpsURLConnection;
  * Created by Afpa on 13/02/2017.
  */
 
-public class FindEvents  extends AsyncTask<String, Void, List<Event>> {
+public class FindEventById extends AsyncTask<String, Void, Event> {
 
-    private final String link = ConfigApi.findevent;
+    private final String link = ConfigApi.findeventbyid;
 
 
     @Override
-    protected List<Event> doInBackground(String... params) {
+    protected Event doInBackground(String... params) {
         StringBuilder sb = new StringBuilder();
         HttpURLConnection urlConnection;
-
-        List<Event> events = new ArrayList<>();
+        Event evt = null;
 
         try {
-            URL url = new URL(link);
+            URL url = new URL(link + params[0]);
             urlConnection = (HttpURLConnection) url.openConnection();
             urlConnection.setReadTimeout(9999);
             urlConnection.setConnectTimeout(9999);
@@ -51,7 +45,7 @@ public class FindEvents  extends AsyncTask<String, Void, List<Event>> {
             urlConnection.setRequestProperty("Accept", "application/json");
             urlConnection.connect();
             if (urlConnection.getResponseCode() == HttpsURLConnection.HTTP_OK) {
-                System.out.println("OK FIND EVENT");
+                System.out.println("OK");
 
                 InputStream in = new BufferedInputStream(urlConnection.getInputStream());
                 BufferedReader br = new BufferedReader(new InputStreamReader(in));
@@ -63,15 +57,12 @@ public class FindEvents  extends AsyncTask<String, Void, List<Event>> {
                 br.close();
 
             } else {
-                System.out.println("PAS OK FIND EVENT");
+                System.out.println("PAS OK");
             }
             urlConnection.disconnect();
 
-            JSONArray jsonArray = new JSONArray(sb.toString());
+            JSONObject jsonObject= new JSONObject(sb.toString());
 
-            for (int i = 0; i < jsonArray.length(); i++) {
-
-                JSONObject jsonObject = jsonArray.getJSONObject(i);
                 int id = jsonObject.getInt("id");
                 String title = jsonObject.getString("title");
                 int user_id = jsonObject.getInt("user_id");
@@ -88,25 +79,24 @@ public class FindEvents  extends AsyncTask<String, Void, List<Event>> {
                 Date eventDate = convertDate(date);
                 Date currentDate = convertDate(creationDate);
 
-                Event evt = new Event(id, title, user_id, eventDate, hour, address, zipcode, city, comment, currentDate, firstname, lastname);
+                evt = new Event(id, title, user_id, eventDate, hour, address, zipcode, city, comment, currentDate, firstname, lastname);
 
-                events.add(evt);
 
-            }
+
 
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return events;
+        return evt;
 
     }
 
     private static Date convertDate(String str) {
         DateFormat formatter = null;
         Date convertedDate = null;
-        formatter = new SimpleDateFormat("yyyy-MM-dd");
+        formatter = new SimpleDateFormat("yyyy-mm-dd");
         try {
-            convertedDate = (Date) formatter.parse(str);
+            convertedDate = formatter.parse(str);
         } catch (ParseException ex) {
 //            Logger.getLogger(Main.class.getName()).log(Level.SEVERE, null, ex);
         }
