+++ app/src/main/java/com/frateinc/yopbooking2/ListEventAdapter.java
@@ -2,12 +2,9 @@ package com.frateinc.yopbooking2;
 
 import android.content.Context;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import android.content.Context;
 import android.content.Intent;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -15,9 +12,8 @@ import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.TableLayout;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import com.frateinc.yopbooking2.models.Event;
+import com.frateinc.yopbooking2.Models.Event;
 
 
 /**
