+++ app/src/main/res/layout/activity_main.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
@@ -8,7 +9,8 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.frateinc.yopbooking2.MainActivity">
+    tools:context="com.frateinc.yopbooking2.MainActivity"
+    android:background="@drawable/bgloginv2">
 
     <ListView
         android:layout_width="match_parent"
@@ -16,35 +18,51 @@
         android:id="@+id/listViewEvents"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
-        android:layout_above="@+id/btnAddEvent"
-        android:layout_alignParentTop="true" />
+        android:layout_below="@+id/txtUserName"
+        android:layout_alignParentBottom="true" />
 
-    <Button
-        android:text="Add event"
+    <TextView
+        android:text="Welcome ###"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/btnAddEvent"
-        android:layout_alignParentBottom="true"
+        android:layout_alignParentTop="true"
+        android:id="@+id/txtUserName"
         android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true" />
+        android:layout_alignParentStart="true"
+        android:textAppearance="@style/TextAppearance.AppCompat.Headline"
+        android:gravity="center"
+
+        android:textColor="@android:color/background_light"
+        android:layout_height="40dp" />
 
     <Button
-        android:text="Login"
+        android:text="@string/switch_user"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/btnLogin"
+        android:layout_above="@+id/listViewEvents"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentEnd="true"
+        android:textAlignment="center"
+        android:gravity="center_vertical|center_horizontal|center"
+        style="@style/Widget.AppCompat.Button.Colored"
+        android:background="@android:color/background_light"
+        android:textColor="@android:color/background_dark"
+        android:paddingLeft="10dp"
+        android:paddingRight="10dp" />
+
+    <android.support.design.widget.FloatingActionButton
+        android:layout_height="wrap_content"
+        android:clickable="true"
+        app:srcCompat="@android:drawable/ic_input_add"
+        android:id="@+id/btnAddEvent"
+        android:layout_width="80sp"
+        app:backgroundTint="?attr/colorPrimary"
+        app:fabSize="normal"
         android:layout_alignParentBottom="true"
         android:layout_alignParentRight="true"
-        android:layout_alignParentEnd="true" />
-
-    <TextView
-        android:text="TextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="13dp"
-        android:id="@+id/txtUserName"
-        android:textSize="24sp" />
+        android:layout_alignParentEnd="true"
+        android:layout_marginRight="10dp"
+        android:layout_marginEnd="10dp"
+        android:layout_marginBottom="11dp" />
 
 </RelativeLayout>
