+++ app/src/main/res/layout/add_event.xml
@@ -110,7 +110,7 @@
             android:gravity="center">
 
             <Button
-                android:text="Add"
+                android:text="@string/add"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:id="@+id/btnAddEvent" />
@@ -118,7 +118,7 @@
     </TableLayout>
 
     <TextView
-        android:text="Add an event"
+        android:text="@string/add_an_event"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/textView13"
