+++ app/src/main/res/values-en/strings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">Yopbooking2</string>
+    <string name="app_name">YopBooking</string>
     <string name="action_settings">Settings</string>
     <string name="organisator">By</string>
     <string name="partyDate"> Date</string>
@@ -9,4 +9,11 @@
     <string name="partyZipcode">Zipcode</string>
     <string name="partyLocation">Location</string>
     <string name="partyName">Party name</string>
+    <string name="add">Add</string>
+    <string name="add_an_event">Add an event</string>
+    <string name="log_in">Log in</string>
+    <string name="please_log_in">Please log in</string>
+    <string name="switch_user">Switch user</string>
+    <string name="welcome_to_yopbooking">Welcome to YopBooking !</string>
+    <string name="register">Register</string>
 </resources>
